requirements.txt
openai
python-dotenv
pydantic
python-multipart
azure-ai-formrecognizer
azure-core
msal
aiofile
fastapi
gunicorn
uvicorn[standard]